Add better card pack structure
New card structure allows each card type and expansion to be parsed for easier construction of component.

Added exampleImages field to Cards type to allow passing in strings to each, which will be passed to an improved ImageContainer.tsx, which will allow multiple reference images of the front/back of provided cards, for easy reference.